pyproject.toml
[tool.poetry]
name = "organization-management-backend"
version = "1.0.0"
description = "A service for organization management"
authors = ["Arnab Adhikari <arnabadhikari93@gmail.com>"]
packages = [{include = "organization_management_backend_service", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.22.0"
requests = "^2.31.0"
websockets = "^11.0"
pydantic = "^2.1.0"
starlette = "0.36.3"
fastapi = "0.109.2"
pydantic-settings = "^2.4.0"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
    files = ["**/*.py"]
    line-length = 160

[tool.isort]
    force_single_line = true
    line_length = 160